split off classes into individual files
