+++ server/tinylicious/src/resources.ts
@@ -1,0 +1,38 @@
+/*!
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import * as services from "@microsoft/fluid-server-services";
+import * as core from "@microsoft/fluid-server-services-core";
+import * as utils from "@microsoft/fluid-server-services-utils";
+import { Provider } from "nconf";
+
+export class TinyliciousResources implements utils.IResources {
+    public webServerFactory: core.IWebServerFactory;
+
+    constructor(
+        public config: Provider,
+        public producer: core.IProducer,
+        public redisConfig: any,
+        public cache: core.ICache,
+        public webSocketLibrary: string,
+        public orderManager: core.IOrdererManager,
+        public tenantManager: core.ITenantManager,
+        public storage: core.IDocumentStorage,
+        public appTenants: core.IAlfredTenant[],
+        public mongoManager: core.MongoManager,
+        public port: any,
+        public documentsCollectionName: string,
+        public metricClientConfig: any,
+        public contentCollection: core.ICollection<any>) {
+
+        this.webServerFactory = new services.SocketIoWebServerFactory(this.redisConfig);
+    }
+
+    public async dispose(): Promise<void> {
+        const producerClosedP = this.producer.close();
+        const mongoClosedP = this.mongoManager.close();
+        await Promise.all([producerClosedP, mongoClosedP]);
+    }
+}
