+++ server/tinylicious/src/resourcesFactory.ts
@@ -1,0 +1,142 @@
+/*!
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import {
+    LocalNodeFactory,
+    LocalOrderManager,
+    NodeManager,
+    ReservationManager,
+} from "@microsoft/fluid-server-memory-orderer";
+import * as services from "@microsoft/fluid-server-services";
+import * as core from "@microsoft/fluid-server-services-core";
+import * as utils from "@microsoft/fluid-server-services-utils";
+import * as bytes from "bytes";
+import { Provider } from "nconf";
+import * as os from "os";
+import * as redis from "redis";
+import { TinyliciousResources } from "./resources";
+import { OrdererManager } from "./ordererManager";
+import { NodeWebSocketServer } from "./webSocketServer";
+
+export class TinyliciousResourcesFactory implements utils.IResourcesFactory<TinyliciousResources> {
+    public async create(config: Provider): Promise<TinyliciousResources> {
+        // Producer used to publish messages
+        const kafkaEndpoint = config.get("kafka:lib:endpoint");
+        const kafkaLibrary = config.get("kafka:lib:name");
+        const kafkaClientId = config.get("alfred:kafkaClientId");
+        const topic = config.get("alfred:topic");
+        const metricClientConfig = config.get("metric");
+        const maxKafkaMessageSize = bytes.parse(config.get("kafka:maxMessageSize"));
+        const producer = services.createProducer(
+            kafkaLibrary,
+            kafkaEndpoint,
+            kafkaClientId,
+            topic,
+            maxKafkaMessageSize);
+        const redisConfig = config.get("redis");
+        const webSocketLibrary = config.get("alfred:webSocketLib");
+        const authEndpoint = config.get("auth:endpoint");
+
+        // Redis connection
+        const redisClient = redis.createClient(redisConfig.port, redisConfig.host);
+        const redisCache = new services.RedisCache(redisClient);
+
+        // Database connection
+        const mongoUrl = config.get("mongo:endpoint") as string;
+        const mongoFactory = new services.MongoDbFactory(mongoUrl);
+        const mongoManager = new core.MongoManager(mongoFactory);
+        const documentsCollectionName = config.get("mongo:collectionNames:documents");
+
+        // create the index on the documents collection
+        const db = await mongoManager.getDatabase();
+        const documentsCollection = db.collection<core.IDocument>(documentsCollectionName);
+        await documentsCollection.createIndex(
+            {
+                documentId: 1,
+                tenantId: 1,
+            },
+            true);
+        const deltasCollectionName = config.get("mongo:collectionNames:deltas");
+        const scribeCollectionName = config.get("mongo:collectionNames:scribeDeltas");
+
+        // foreman agent uploader does not run locally.
+        // TODO: Make agent uploader run locally.
+        const foremanConfig = config.get("foreman");
+        const taskMessageSender = services.createMessageSender(config.get("rabbitmq"), foremanConfig);
+        await taskMessageSender.initialize();
+
+        const nodeCollectionName = config.get("mongo:collectionNames:nodes");
+        const nodeManager = new NodeManager(mongoManager, nodeCollectionName);
+        // this.nodeTracker.on("invalidate", (id) => this.emit("invalidate", id));
+        const reservationManager = new ReservationManager(
+            nodeManager,
+            mongoManager,
+            config.get("mongo:collectionNames:reservations"));
+
+        const tenantManager = new services.TenantManager(authEndpoint, config.get("worker:blobStorageUrl"));
+
+        const databaseManager = new core.MongoDatabaseManager(
+            mongoManager,
+            nodeCollectionName,
+            documentsCollectionName,
+            deltasCollectionName,
+            scribeCollectionName);
+
+        const storage = new services.DocumentStorage(databaseManager, tenantManager, producer);
+
+        const maxSendMessageSize = bytes.parse(config.get("alfred:maxMessageSize"));
+
+        const contentCollection = db.collection("content");
+        await contentCollection.createIndex(
+            {
+                documentId: 1,
+                sequenceNumber: 1,
+                tenantId: 1,
+            },
+            false);
+
+        const address = `${await utils.getHostIp()}:4000`;
+        const nodeFactory = new LocalNodeFactory(
+            os.hostname(),
+            address,
+            storage,
+            databaseManager,
+            60000,
+            () => new NodeWebSocketServer(4000),
+            taskMessageSender,
+            tenantManager,
+            foremanConfig.permissions,
+            maxSendMessageSize);
+        const localOrderManager = new LocalOrderManager(nodeFactory, reservationManager);
+        const serverUrl = config.get("worker:serverUrl");
+
+        const orderManager = new OrdererManager(
+            serverUrl,
+            tenantManager,
+            localOrderManager);
+
+        // Tenants attached to the apps this service exposes
+        const appTenants = config.get("alfred:tenants") as { id: string, key: string }[];
+
+        // This wanst to create stuff
+        const port = utils.normalizePort(process.env.PORT || "3000");
+
+        return new TinyliciousResources(
+            config,
+            producer,
+            redisConfig,
+            redisCache,
+            webSocketLibrary,
+            orderManager,
+            tenantManager,
+            storage,
+            appTenants,
+            mongoManager,
+            port,
+            documentsCollectionName,
+            metricClientConfig,
+            contentCollection);
+    }
+}
